# optimise versions methods