- refactor(qrcode): optimise versions related methods
- refactor(qrcode): optimise loadVersion again, reduce function call in limitation scope