@yeqown 2021-11-29
```sh
$ benchstat old.txt new.txt
name                   old time/op  new time/op  delta
_loadVersion_top-4     55.0ns ±32%   4.8ns ± 7%  -91.29%  (p=0.008 n=5+5)
_loadVersion_waist-4   47.6ns ± 9%   3.4ns ± 8%  -92.93%  (p=0.008 n=5+5)
_loadVersion_bottom-4  57.0ns ± 8%   7.5ns ± 5%  -86.89%  (p=0.008 n=5+5)
```

basis: in a limited scope, we have no need to use binary search, just shrink the range to search, reduce a function all can speed up this.